Merge branch testing to develop

See merge request fsel/api/fsel_src!35

+++ Services/Fsel.Identity/Fsel.Identity.Domain/Entities/StudentRanking.cs
@@ -11,7 +11,7 @@ namespace Fsel.Identity.Domain.Entities
 
         public int DailyStreak { get; set; }
 
-        public int TotalScore { get; set; }
+        public double TotalScore { get; set; }
 
         public int PositionChange { get; set; }
